fix updating token values
