+++ app/state/applets/applets.thunks.js
@@ -351,8 +351,8 @@ export const downloadApplets = (onAppletsDownloaded = null, keys = null) => asyn
         dispatch(setCumulativeActivities(cumulativeActivities));
         dispatch(setClosedEvents(finishedEvents));
 
-        // await storeData('ml_applets', transformedApplets);
-        // await storeData('ml_responses', responses);
+        await storeData('ml_applets', transformedApplets);
+        await storeData('ml_responses', responses);
 
         if (scheduleUpdated) {
           dispatch(setScheduleUpdated(true));
